Links Component Types
- Moved types to their own files (and a sep type folder within links component folder)
- Changed a few type names to make their purpose clearer
