+++ packages/frontend/components/links/types/deleteLink.d.ts
@@ -1,0 +1,4 @@
+export interface IDeleteResponse {
+  success: boolean;
+  error?: string;
+}
