+++ packages/frontend/components/links/types/getLinks.d.ts
@@ -6,27 +6,6 @@ export interface ILink {
   _id: string;
 }
 
-export interface IShortenAPI {
-  longUrl: string;
-  shortUrl?: string;
-}
-
-export interface IShortenResponse {
-  success: boolean;
-  shortUrl: string;
-  error?: string;
-}
-export interface IGetUrlResponse {
-  success: boolean;
-  longUrl: string;
-  error?: string;
-}
-
-export interface IDeleteResponse {
-  success: boolean;
-  error?: string;
-}
-
 export interface IGetLinksAxiosResponse {
   success: boolean;
   links: Array<ILink>;
@@ -39,5 +18,3 @@ export interface IGetLinksResponse {
   isLoading: boolean;
   mutate: (data?: Data | Promise<Data> | mutateCallback<Data>, shouldRevalidate?: boolean) => Promise<Data | undefined>;
 }
-
-export type LinksFn = (LinkData: IShortenAPI) => Promise<IShortenResponse>;
