+++ packages/frontend/components/links/types/getLongUrl.d.ts
@@ -1,0 +1,5 @@
+export interface IGetLongUrlResponse {
+  success: boolean;
+  longUrl: string;
+  error?: string;
+}
